Switch back to naive datetimes
Also removed unnecessary time_changed and time_created columns from all
tables except the draft_mod table, since that's where they were needed.

In the future, we can have a better wrapper that stores a timezone-naive
`TIMESTAMP` in the DB, but returns a timezone-aware `datetime`, but for
now this works.